removed I2C, wire and MMA8453Q libraries
MMA was used in NB v1 but not present in V2.1. Therefore, the lib for
that as well as I2C.h and wire.h libs which are used for comms with it
are not needed and can be removed to gain code space. required code
space reduced from 29k to 26k.
!!!!will only compile with V2.1 macro enabled in ninja.h!!!