Adding status LED Blinking feature
Status LED color change message can now be extended to set a period and
duty cycle for blinking the same LED. The message format is
{"DEVICE":[{"G":"0","V":0,"D":999,"DA":"DDPPPPCCCCCC"}]}
where DD is the duty cycle, PPPP is the period in ms and CCCCCC is the
color. Setting the duty cycle to 0 with a non zero period will disable
blinking.  If duty cycle and period values are not included or if period
is zero then only color will change. So the previous color change
message still functions as intended.